+++ src/net/sourceforge/kolmafia/request/CafeRequest.java
@@ -72,7 +72,7 @@ public class CafeRequest extends GenericRequest {
     this.addFormField("whichitem", String.valueOf(itemId));
   }
 
-  public static int discountedPrice(int price) {
+  public static long discountedPrice(long price) {
     int count = LARP.getCount(KoLConstants.inventory) + LARP.getCount(KoLConstants.closet);
 
     if (count > 0) {
@@ -215,7 +215,7 @@ public class CafeRequest extends GenericRequest {
     return true;
   }
 
-  public static final void registerItemUsage(final String itemName, int price) {
+  public static final void registerItemUsage(final String itemName, long price) {
     int inebriety = ConsumablesDatabase.getInebriety(itemName);
     String consume = inebriety > 0 ? "drink" : "eat";
